+++ CHANGELOG.md
@@ -22,6 +22,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   `OTEL_EXPORTER_OTLP_TRACES_COMPRESSION`,
   `OTEL_EXPORTER_OTLP_METRICS_COMPRESSION` and
   `OTEL_EXPORTER_OTLP_LOGS_COMPRESSION`.
+- Support for Database OpenTelemetry Tracing for [Oracle.ManagedDataAccess.Core](https://www.nuget.org/packages/Oracle.ManagedDataAccess.Core)
+  and [Oracle.ManagedDataAccess](https://www.nuget.org/packages/Oracle.ManagedDataAccess).
 
 ### Changed
 
@@ -39,7 +41,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   - `OpenTelemetry.Instrumentation.SqlClient` from `1.15.2` to `1.16.0`,
   - `OpenTelemetry.Instrumentation.StackExchangeRedis` from `1.15.1-beta.2` to `1.16.0-beta.1`,
   - `OpenTelemetry.Instrumentation.Wcf` from `1.15.1-beta.2` to `1.16.0-beta.1`,
-  - `OpenTelemetry.OpAmp.Client` from `0.3.0-alpha.1` to `0.4.0-alpha.1`.
+  - `OpenTelemetry.OpAmp.Client` from `0.3.0-alpha.1` to `0.5.0-alpha.1`.
 - .NET only, following packages updated
   - `OpenTelemetry.Instrumentation.AspNetCore` from `1.15.2` to `1.16.0`,
   - `OpenTelemetry.Instrumentation.EntityFrameworkCore`